clang: Remaining changes
Those are a bit less obvious than those I pushed directly to master.
All those changes were required to build LXC under clang here.

With this, gcc can be replaced by clang to build LXC so long as you're
not using the python3 binding (as python extensions can't be built under
clang at the moment).

For reference, the clang output for those is: http://paste.ubuntu.com/6292460/

Acked-by: Serge E. Hallyn <serge.hallyn@ubuntu.com>

+++ src/lxc/caps.c
@@ -204,7 +204,7 @@ static int _real_caps_last_cap(void)
 			buf[n] = '\0';
 			result = strtol(buf, &ptr, 10);
 			if (!ptr || (*ptr != '\0' && *ptr != '\n') ||
-			    result == LONG_MIN || result == LONG_MAX)
+			    result == INT_MIN || result == INT_MAX)
 				result = -1;
 		}
